+++ src/borg/chunkers/buzhash64.pyx
@@ -237,8 +237,7 @@ cdef class ChunkerBuzHash64:
                 raise Exception("chunkifier byte count mismatch")
 
         while self.remaining < min_size + window_size + 1 and not self.eof:  # see assert in Chunker init
-            if not self.fill():
-                return None
+            self.fill()
 
         # Here we either are at eof...
         if self.eof:
@@ -294,8 +293,7 @@ cdef class ChunkerBuzHash64:
             self.remaining -= did_bytes
 
             if self.remaining <= window_size:
-                if not self.fill():
-                    return None
+                self.fill()
 
         if self.remaining <= window_size:
             self.position += self.remaining